Reworks existing DB check, always runs `goose`.
Prior to this commit an optimization was present in `test/create_db.sh`
that would `exit 0` if the `boulder_sa_integration` database existed.
This early exit meant that after the first invocation of `create_db.sh`
no further `goose` migrations would be applied unless the operator
dropped their databases or edited the script.

This commit reworks the existing DB optimization so that it only skips
the `CREATE DATABASE` statements and allows `goose` to try and apply
migrations.

This doesn't result in significantly longer start up times because Goose
is smart enough to know when no migrations are required and outputs
something similar to:
  `goose: no migrations to run. current version: 20160602142227`

This should address #2174.

+++ test/create_db.sh
@@ -9,11 +9,6 @@ if [[ $MYSQL_CONTAINER ]]; then
 	dbconn="-u root -h boulder-mysql --port 3306"
 fi
 
-if mysql $dbconn -e 'show databases;' | grep boulder_sa_integration > /dev/null; then
-  echo "Databases already created."
-  exit 0
-fi
-
 # MariaDB sets the default binlog_format to STATEMENT,
 # which causes warnings that fail tests. Instead set it
 # to the format we use in production, MIXED.
@@ -22,11 +17,16 @@ mysql $dbconn -e "SET GLOBAL binlog_format = 'MIXED';"
 for dbenv in $DBENVS; do
   (
   db="boulder_sa_${dbenv}"
-  create_script="drop database if exists \`${db}\`; create database if not exists \`${db}\`;"
 
-  mysql $dbconn -e "$create_script" || die "unable to create ${db}"
+  if mysql $dbconn -e 'show databases;' | grep $db > /dev/null; then
+    echo "Database $db already exists - skipping create"
+  else
+    create_script="drop database if exists \`${db}\`; create database if not exists \`${db}\`;"
 
-  echo "created empty ${db} database"
+    mysql $dbconn -e "$create_script" || die "unable to create ${db}"
+
+    echo "created empty ${db} database"
+  fi
 
   goose -path=./sa/_db/ -env=$dbenv up || die "unable to migrate ${db}"
   echo "migrated ${db} database"
